+++ public/sw.js
@@ -1,4 +1,4 @@
-const SW_VERSION='14.8'
+const SW_VERSION='14.9'
 self.addEventListener('install',event=>{self.skipWaiting()})
 self.addEventListener('activate',event=>{
   event.waitUntil((async()=>{